+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- **The pending comment links to the run logs.** Just the same as the completed report comment.
+
 ## [0.8.0] - 2026-08-21
 
 ### Added
@@ -174,6 +180,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 Initial release.
 
+[Unreleased]: https://github.com/ritten-org/Ritten/compare/v0.8.0...HEAD
 [0.8.0]: https://github.com/ritten-org/Ritten/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/ritten-org/Ritten/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/ritten-org/Ritten/compare/v0.5.0...v0.6.0